server/requirements.txt
aiosqlite==0.17.0
bcrypt==4.0.0
databases==0.6.1
fastapi==0.85.0
passlib==1.7.4
pydantic==1.10.2
python-jose==3.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
uvicorn==0.18.3